bump dwn-sdk-js and resolve LevelDB locking issues
