+++ .gitignore
@@ -13,3 +13,5 @@ node_modules
 tmp
 dwn.db
 coverage
+
+RESOLVERCACHE
