+++ package.json
@@ -26,7 +26,7 @@
     "url": "https://github.com/TBD54566975/dwn-server/issues"
   },
   "dependencies": {
-    "@tbd54566975/dwn-sdk-js": "0.2.16",
+    "@tbd54566975/dwn-sdk-js": "0.2.17",
     "@tbd54566975/dwn-sql-store": "0.2.9",
     "better-sqlite3": "^8.5.0",
     "body-parser": "^1.20.2",
@@ -58,6 +58,7 @@
     "@types/ws": "8.5.4",
     "@typescript-eslint/eslint-plugin": "5.59.0",
     "@typescript-eslint/parser": "5.59.0",
+    "@web5/dids": "0.4.0",
     "c8": "8.0.1",
     "chai": "4.3.6",
     "chai-as-promised": "7.1.1",
